requirements.txt
httpx>=0.24.0
mcp-server>=0.1.0
mcp-client>=0.1.0 